Completed assignment. issues with code

+++ src/main/java/com/mycompany/csc229_211review_lab02hw/Person.java
@@ -8,10 +8,10 @@ public abstract class Person {
 	
 	private String name;
 	private String address;
-	private short age;
+	private int age;
 	
 	// constructor that takes only two paras
-	public Person(String name, short age) {
+	public Person(String name, int age) {
 		super();
 		this.name = name;
 		this.age = age;
@@ -30,11 +30,11 @@ public abstract class Person {
 	public abstract void setAddress(String address);
 	
 	
-	public short getAge() {
+	public int getAge() {
 		return age;
 	}
 	
-	public void setAge(short age) {
+	public void setAge(int age) {
 		this.age = age;
 	}
